+++ src/vatensor/linalg.h
@@ -6,6 +6,7 @@
 
 namespace va {
     void dot(VArrayTarget target, const VArray& a, const VArray& b, const Axes& axes);
+    void matmul(VArrayTarget target, const VArray& a, const VArray& b);
 }
 
 #endif //LINALG_H